+++ packages/terra-functional-testing/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Added
+  * Added describeTests helper to filter tests by form factors, locales, or themes
+
 ## 1.0.1 - (March 1, 2021)
 
 * Fixed